Created display_init.c file and moved the main init of the display from the main.c file to there. Also made start.c as a file to start everything up
