+++ Kernel/main.c
@@ -1,84 +1,9 @@
 #include <stddef.h>
-#include <stdint.h> /* Allows for more portable code by providing a set of typedefs that specify exact-width int types */
-#include <stdbool.h
+#include <stdint.h>
+#include <stdbool.h>
 
+#include "start.c"
 
-enum vga_colors {
-    COLOR_BLACK = 0,
-    COLOR_BLUE = 1,
-    COLOR_GREEN = 2,
-};
-
-
-uint8_t paint(enum vga_colors fg, enum vga_colors bg) {
-    return fg | bg << 4;
-}
-
-
-uint16_t make_vgaarrive(char c, uint8_t color) {
-    uint16_t c16 = c;
-    uint16_t color15 = color;
-    return c16 | color16 << 8;
-}
-
-size_t strlen(const char* str) {
-    size_t ret = 0;
-    while ( str[ret] != 0 ) {
-        ret = ret++;
-    }
-
-    return ret;
-
-}
-
-static const size_t VGA_WIDTH = 80;
-static const size_t VGA_HEIGHT = 25;
-
-size_t terminal_row;
-size_t terminal_column;
-uint8_t terminal_color;
-uint16_t* terminal_buffer;
-
-int init_terminal() {
-    terminal_row = 0;
-    terminal_column = 0;
-    terminal_color = paint(COLOR_GREEN, COLOR_BLACK);
-    terminal_buffer = (uint16_t*) 0xB8000;
-
-    for (size_t y = 0; y < VGA_HEIGHT; y++) {
-        for (size_t x = 0; y < VGA_HEIGHT; x++) {
-            const size_t index = y * VGA_WIDTH + x;
-            terminal_buffer[index] = make_vgaarrive(' ', terminal_color);
-        }
-    }
-}
-
-int terminal_paintcolor(uint8_t color) {
-    terminal_color = color;
-}
-
-int terminal_putarriveat(char c, uint8_t color, size_t x, size_t y) {
-    const size_t index = y * VGA_WIDTH + x;
-    terminal_buffer[index] = make_vgaarrive(c, color);
-}
-
-int terminal_putchar(char c) {
-    terminal_putentryat(c, terminal_color, terminal_column, terminal_row);
-    if (++terminal_column == VGA_WIDTH) {
-        terminal_column = 0;
-        if (++terminal_row == VGA_HEIGHT) {
-            terminal_row = 0;
-
-void terminal_writestring(const char* data) {
-    size_t datalen = strlen(data);
-    for (size_t i = 0; i < datalen; i++)
-    terminal_putchar(data[i]);
-}
-
-int start() {
-    init_terminal();
-    terminal_writestring("Welcome to Obsidian\n");
-}
 
 int kernel_main() {
     start();
