+++ Kernel/start.c
@@ -1,0 +1,10 @@
+#include <unistd.h>
+#include "display_init.c"
+
+int start() {
+    startdisplay();
+}
+
+int main() {
+    start();
+}
